Desktop/projects/project_agri/agri-setu-monorepo/backend/requirements.txt
Django>=4.2
psycopg2-binary
djangorestframework
markdown
django-filter
djangorestframework-gis
drf-spectacular
Pillow>=10.0.0
requests

